# internal/server/swagger/swagger.yaml
basePath: /
...
host: localhost:1324
info:
  contact: {}
  description: KNOV API \n http://localhost:1324
  title: Knov API
  version: "1.0"
paths:
...
  /api/logs/download:
    get:
      description: Downloads the raw contents of a single log file as plain text
      parameters:
      - description: 'log file name (default: the active app.log)'
        in: query
        name: name
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: log file contents
          schema:
            type: file
        "404":
          description: file logging not enabled
          schema:
            type: string
      summary: Download a log file
      tags:
      - system
  /api/logs/file:
    get:
      description: Returns lines from a log file for the file-view tab. name selects
        a single key's log file (e.g. file-sync.log), name=all merges every current
        key's log into one chronologically sorted table, or name is omitted for the
        active app.log. chunk/limit/offset page through a single file (not supported
        for name=all).
      parameters:
      - description: log file name, or 'all' to merge every key's log
        in: query
        name: name
        type: string
      - description: max lines/entries to return (default 1000)
        in: query
        name: limit
        type: integer
      - description: lines to skip from the end, for paging a single file
        in: query
        name: offset
        type: integer
      - description: return only the appended fragment, without the surrounding container
        in: query
        name: chunk
        type: boolean
      produces:
      - text/html
      responses:
        "200":
          description: log lines HTML
          schema:
            type: string
        "404":
          description: file logging not enabled
          schema:
            type: string
      summary: Get log file contents
      tags:
      - system
...
  /api/system/cache:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: Removes all cache entries, forcing a rebuild on next access
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: cache invalidated
          schema:
            type: string
        "500":
          description: failed to invalidate cache
          schema:
            type: string
      summary: Invalidate cache
      tags:
      - system
...
  /api/themes/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Set new theme via form parameter
      parameters:
      - description: Theme name to set
        in: formData
        name: theme
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "303":
          description: Redirect to settings page
      summary: Set theme
      tags:
      - themes
  /api/themes/{themeName}/settings:
    get:
      description: Get all settings for a specific theme
      parameters:
      - description: Theme name
        in: path
        name: themeName
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: Theme settings object
          schema:
            type: object
      summary: Get theme settings
      tags:
      - themes
  /api/themes/{themeName}/settings/{settingKey}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update a specific setting for a theme
      parameters:
      - description: Theme name
        in: path
        name: themeName
        required: true
        type: string
      - description: Setting key to update
        in: path
        name: settingKey
        required: true
        type: string
      - description: Setting value
        in: formData
        name: value
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: Setting updated successfully
      summary: Update theme setting
      tags:
      - themes
  /api/themes/settings:
    get:
      description: Get all theme settings as HTML form elements
      produces:
      - text/html
      responses:
        "200":
          description: HTML form elements
          schema:
            type: string
      summary: Get theme settings form
      tags:
      - themes
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Update a specific setting for the current theme
      parameters:
      - description: Setting key
        in: formData
        name: key
        required: true
        type: string
      - description: Setting value
        in: formData
        name: value
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: Setting updated successfully
      summary: Update theme setting
      tags:
      - themes
swagger: "2.0"